Makefile: Allow running `make test` before `make install`
The current invocation of ./cyclone in the test target of the Makefile
assumes that cyclone is already installed globally to find libraries
required for the test programs. However, it may be desirable to execute
the tests before installing Cyclone. For instance, package recipes for
Linux distributions usually run tests before the install step.